refactor(planning): inline ticket-decomposer engine into epic-plan-decompose and retire the file (resolves #1447)
